[#932] Apply ruff auto-fixes and fix a curated set of real lint findings
Ran `ruff check --fix` (safe autofixes: import sorting, redundant
parens, collapsed elif/else, implicit string concatenation, etc.) and
`ruff format` across the repo, then manually triaged the remaining
small, non-cosmetic categories in `DisplayCAL` (the large ones -
missing type annotations, missing docstrings, naming conventions,
`C901` complexity - are left alone as out of scope for this pass):

- `display_cal.py`: removed a duplicate `get_argyll_util` import
  (`F811`).
- `config.py`: removed an unreachable `elif orig_name in ():` branch
  (empty tuple, always `False`); combined a lock file/`FileLock`
  nested `with` into one statement; replaced list concatenation with
  unpacking in the two `os.execv()` relaunch calls.
- `edid.py`: removed a dead `edid = None` assignment in
  `get_edid_windows()`; collapsed a nested `if` in the xrandr parser;
  silenced a false-positive ambiguous-unicode warning on a regex that
  intentionally matches `×` in macOS resolution strings.
- `icc_profile.py`: removed an unused `win32api` import (kept as a
  `noqa`'d side-effect import, its presence in `sys.modules` is
  checked elsewhere); converted two `if`/`else` blocks to ternaries;
  restored (and documented via `# noqa`/`pyproject.toml`) the
  `round()` calls in the `uInt*Number_tohex()` helpers that a first
  pass removed as "redundant" based on the `num: int` type hint alone,
  the hint doesn't match real call sites, which pass floats, and
  removing `round()` crashed `struct.pack()` at runtime.
- `worker.py`: merged two `==` comparisons into one `in (...)` check.
- `ui/update_check_window.py`: annotated a mutable class-level dict
  default with `ClassVar`.
- `ui/plot/ccxx_data.py`, `util_os.py`: removed two more dead/unused
  local variables.
- `ui/plot/ccxx.py`: replaced stray U+2009/U+2002 thin/en-space
  characters in a legend label with plain ASCII spaces (no rationale
  found in history for the original choice).
- `calibration_file.py`, `ui/measurement_report.py`,
  `ui/untethered_window.py`, `profile_name.py`: replaced four
  `try`/`except`/`pass` blocks with `contextlib.suppress(...)`.
- `eecolor_to_madvr.py`: converted several `%`-style `safe_print()`
  calls to f-strings.
- `wx_display_adjustment_frame.py`, `wx_fixes.py`, `worker.py`:
  replaced three more list-concatenation patterns with unpacking.
- `pyproject.toml`: ignore `S311` (non-cryptographic `random` use is
  expected here), `RUF057` (flags `round()` calls the annotations
  don't actually justify removing) and `PYI041` (the codebase already
  uses explicit `int | float` unions deliberately in several spots).

Also fixed two Qt tests (`test_ui_spyder2_enable.py`,
`test_ui_patterngenerator_setup.py`) that reached through the module
under test for `QMessageBox` as a monkeypatch anchor; that import was
itself dead code (both modules go through the `message_box` wrapper),
so removing it broke the tests. Now they import `QMessageBox` directly
from `qtpy.QtWidgets`.

Full test suite (2576 passed, 21 skipped) run serially with no
regressions.

+++ DisplayCAL/wx_visual_whitepoint_editor.py
@@ -2701,7 +2701,7 @@ class VisualWhitepointEditor(wx.Frame):
             ]
             pos[0] += self.mainPanel.Position[0]
             pos[1] -= (self.Size[0] - self.ClientSize[0]) / 2
-            pos = [int(round(coord)) for coord in pos]
+            pos = [round(coord) for coord in pos]
             self._mgr.GetPane("mainPanel").FloatingPosition(pos).CloseButton(True)
             wx.CallAfter(self.area_handler)
         else:
